Move internal class to internal package
